Moved generic functions for generating GOOL from Import to GenerateGOOL
